requirements.txt
pandas==1.5.0
psycopg2==2.9.5
psycopg2_binary==2.9.5
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.44
